+++ lib/src/components/dialogs/MoreInfoDlg.tsx
