+++ lib/src/declare.d.ts
@@ -1,2 +1,1 @@
 declare module 'stockholm-js'
-declare module 'normalize-wheel'
